Remove exec file form Janus - Fixes #13
