--- CHANGELOG.md
@@ -1,9 +1,0 @@
-# [0.0.4] - 2016-03-08
-- Feature/fake - [PR Link](https://github.com/thebeansgroup/januscli/pull/11)
- - Local issue - [Fixes #10](https://github.com/thebeansgroup/januscli/issues/10) - Local issue - [Fixes #10](https://github.com/thebeansgroup/januscli/issues/10)# [0.0.3] - 2016-03-08
-- Release: 0.0.2 - [PR Link](https://github.com/thebeansgroup/januscli/pull/8)
- - Test Bug Issue Title - [Fixes #6](https://github.com/thebeansgroup/januscli/issues/6)# [0.0.2] - 2016-03-08
-- Initial test environment for JanusCli Release - [PR Link](https://github.com/thebeansgroup/januscli/pull/7)
- - Test Bug Issue Title - [Fixes #6](https://github.com/thebeansgroup/januscli/issues/6)# [0.0.1] - 2016-03-03
-- OLD Initial test environment for JanusCli Release - [PR Link](https://github.com/thebeansgroup/januscli/pull/7)
- - OLD Test Bug Issue Title - [Fixes #6](https://github.com/thebeansgroup/januscli/issues/6)## Next
