--- bin/index.js
@@ -1,4 +1,0 @@
-#!/usr/bin/env node
-
-require('../es5');
-
