+++ package.json
@@ -1,11 +1,8 @@
 {
   "name": "januscli",
   "version": "0.1.0",
-  "description": "Development toolbelt",
-  "main": "es5/janus.js",
-  "bin": {
-    "janus": "bin/index.js"
-  },
+  "description": "Development toolbelt framework",
+  "main": "es5/index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "start": "babel --watch --presets es2015 lib -d es5",
